l3 is fixed
