--- lesson3/task1_2_3/exceptions/FolderCreationException.java
@@ -1,7 +1,0 @@
-package lesson3.task1_2_3.exceptions;
-
-public class FolderCreationException extends Exception{
-    public FolderCreationException(String message) {
-        super(message);
-    }
-}
