--- lesson3/task1_2_3/exceptions/UnknownOsException.java
@@ -1,7 +1,0 @@
-package lesson3.task1_2_3.exceptions;
-
-public class UnknownOsException extends Exception {
-    public UnknownOsException(String message) {
-        super(message);
-    }
-}
