+++ lesson3/tasks/exceptions/FileCreationException.java
@@ -1,4 +1,4 @@
-package lesson3.task1_2_3.exceptions;
+package lesson3.tasks.exceptions;
 
 public class FileCreationException extends Exception {
     public FileCreationException(String s) {
